Setup  marko-web/middleware.js & set index.js to empty
